Set org.opencontainers.image.source label in Dockerfile
As documented in the OCI Image Format,
org.opencontainers.image.source[1] identifies an image's source
repository. This is purely for documentation purposes. It does however
help tools such as Renovate[2] to find the changelogs when a new Traggo
version is released. The changelogs would then be included in the PR
Renovate creates.

[1]: https://github.com/opencontainers/image-spec/blob/5325ec48851022d6ded604199a3566254e72842a/annotations.md#L24
[2]: https://docs.renovatebot.com/modules/datasource/docker/#description