Add setting to use browser native date-time picker (#204)
* style: autoformat

Running `make format` on a clean copy of this repository made these
changes

* feat(datetimeinputstyle): add `DatetimeInputStyle` setting

This boilerplate commit adds a `DatetimeInputStyle` field to user
settings on both the client and server. It can represent either "Fancy"
(this is the input element currently present in Traggo) or "Standard"
(unimplemented, the <input type="datetime-local"> HTML element).
A later commit will actually do something with this commit.

* feat(datetimeinputstyle): add `<input type=datetime-local>` time picker option

* fix(datetimeinputstyle): fix date formatting

* refactor(datetimeinputstyle): rename `Standard` to `Native`

* refactor(datetimeinputstyle): change snackbar message

* refactor(datetimeinputstyle): replace `DatetimeInput` with `DateTimeInput`